api/requirements.txt
Django==2.2.4
django-extensions==2.2.1
django-filter==2.2.0
djangorestframework==3.10.2
Markdown==3.1.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
Werkzeug==0.15.5
whitenoise==4.1.3
